refactor(chatbot): separating chatbot and FaQ files in different folders
